Board and alight panels and controls done. We'll need an info panel with number of passengers that weren't able to get onboard, and maybe some other things.
Board passengers is not quite working. Because I have to update the
displays of 
the carriages when passengers are boarded, I'm grabbing all of the old
panels, getting their toStrings again and updating them, then adding
these to an array,
removing the old panels and then adding them back on. Except that the
new panels are not being added. I have to go out now, will be back home
around 6 or 7.